Move ConnectionManager tests with the other tests for that class
Remove the duplicate stub code and move the tests for ConnectionManager
code with the rest of that class' tests.
